Test improvements: only load `.env.test` and don't show logging for passing tests (it's too noisy) (#56)
Not sure in retrospect why I made it load `.env` in tests. Anyway, we don't want to dirty our test setup with per-developer config,
so only load `.env.test` to make sure we have a reproducible setup.

Finally, ask Vitest to suppress logs for passing tests. Makes the test output a lot cleaner and means the logging reflects your failing tests, so it's easier to understand.
(This is how Cargo does it in Rust and it seems to work fine there.)

Reviewed-on: https://git.emunest.net/Captain_Orb/MolCI/pulls/56
Reviewed-by: Captain_Orb <tehorb13@gmail.com>
Co-authored-by: Olivier 'reivilibre <git.contact@librepush.net>
Co-committed-by: Olivier 'reivilibre <git.contact@librepush.net>